requirements.txt
python-telegram-bot==20.3
firebase-admin==6.3.0
requests
qrcode
pillow
pytonlib==0.1.1
